移除控件dock特性,因为项目中基本上用不到,控件OnKillFocus函数新增一个参数 该参数是下一个具有焦点的控件,新增OnRect消息事件 ,新增对窗口最大化最小化缩放尺寸 优化部分代码

+++ EzUI/sources/EzUI.cpp
@@ -80,9 +80,9 @@ namespace EzUI {
 			mt.Physical.Width = dm.dmPelsWidth;//物理宽
 			mt.Physical.Height = dm.dmPelsHeight;//物理高
 			//计算缩放
-			mt.Scale = ((double)mt.Physical.Height / (double)mt.Rect.Height);
+			mt.Scale = ((float)mt.Physical.Height / (float)mt.Rect.Height);
 			//显示器fps
-			mt.FPS = dm.dmDisplayFrequency;
+			mt.FPS = (float)dm.dmDisplayFrequency;
 			monitors->push_back(mt);
 			return TRUE;
 			}, LPARAM(outMonitorInfo));
